Polish, fixes, organizing; View rewrited as a Control bag
