+++ GridOS/GridOS.projitems
@@ -9,15 +9,27 @@
     <Import_RootNamespace>GridOS</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Data Objects\ConfigurationDump.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data Objects\ObservableT.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data Objects\ObservableUpdateFrequency.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Enums\Enums.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Built-In Menu\HelpMenu.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessing.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\Controls\Breadcrumb.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\Controls\MenuContentBuilder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\Controls\NavigationFrame.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\Controls\DisplayHeader.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\Controls\ScrollableFrame.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\Controls\_ControlBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\DisplayController.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\DisplayView.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\DisplayViewModel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\DisplayElementMenuHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessors\WordWrap_BreakPresearchStrategy.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessors\PadAllLines.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessors\LineInfoExtractor.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessors\AddPrefix.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessors\AddSuffix.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Logic\TextProcessors\_TextProcessorBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Model\DisplayElement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Model\DisplayCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Main Components\Display Model\DisplayGroup.cs" />
